Rename and move api into geo
